Fixing Himeno's validation
After some investigations, turns out that the result is too noisy when
using float due to standard floating point imprecision, not anything
horrible the compilers were making or special x86_64/AArch64 instruction
patterns. So, changing the matrices to double fixes all of the problems.

However, it also doubles the BSS size, which then in turn needs
-mcmodel=large on x86_64 for -DLARGE. AArch64 doesn't need because its
default relocation already spans across the whole 32-bit range.

Some investigation was needed to make sure that wouldn't make a huge
difference in the results. Moving to double has a 25% impact in
"performance", but it also makes it the same source for all
architectures and relocation sizes, so it's a fairer comparison.